+++ compiler-core/src/language_server/engine.rs
@@ -109,8 +109,7 @@ where
         self.progress_reporter.compilation_finished();
 
         let modules = result?;
-        self.modules_compiled_since_last_feedback
-            .extend(modules.into_iter());
+        self.modules_compiled_since_last_feedback.extend(modules);
 
         Ok(())
     }
@@ -205,11 +204,9 @@ where
 
                 Located::FunctionBody(_) => Some(this.completion_values(module)),
 
-                Located::ModuleStatement(
-                    Definition::ExternalFunction(_)
-                    | Definition::TypeAlias(_)
-                    | Definition::CustomType(_),
-                ) => Some(this.completion_types(module)),
+                Located::ModuleStatement(Definition::TypeAlias(_) | Definition::CustomType(_)) => {
+                    Some(this.completion_types(module))
+                }
 
                 Located::ModuleStatement(Definition::Import(_) | Definition::ModuleConstant(_)) => {
                     None